+++ evidence/reactor.go
@@ -1,11 +1,15 @@
 package evidence
 
 import (
+	"fmt"
 	"time"
 
+	"github.com/gogo/protobuf/proto"
 	clist "github.com/tendermint/tendermint/libs/clist"
 	"github.com/tendermint/tendermint/libs/log"
 	"github.com/tendermint/tendermint/p2p"
+	ep "github.com/tendermint/tendermint/proto/evidence"
+	tmproto "github.com/tendermint/tendermint/proto/types"
 	"github.com/tendermint/tendermint/types"
 )
 
@@ -208,4 +212,48 @@ func (evR Reactor) checkSendEvidenceMessage(
 // PeerState describes the state of a peer.
 type PeerState interface {
 	GetHeight() int64
+}
+
+// encodemsg takes a array of evidence
+// returns the byte encoding of the List Message
+func encodeMsg(evis []types.Evidence) ([]byte, error) {
+	evi := make([]tmproto.Evidence, len(evis))
+	for i := 0; i < len(evis); i++ {
+		le := evis[i]
+		ev, err := types.EvidenceToProto(le)
+		if err != nil {
+			return nil, err
+		}
+		evi[i] = *ev
+	}
+
+	epl := ep.List{
+		Evidence: evi,
+	}
+
+	return proto.Marshal(&epl)
+}
+
+// decodemsg takes an array of bytes
+// returns an array of evidence
+func decodeMsg(bz []byte) (evis []types.Evidence, err error) {
+	lm := ep.List{}
+	proto.Unmarshal(bz, &lm)
+
+	evis = make([]types.Evidence, len(lm.Evidence))
+	for i := 0; i < len(lm.Evidence); i++ {
+		ev, err := types.EvidenceFromProto(&lm.Evidence[i])
+		if err != nil {
+			return nil, err
+		}
+		evis[i] = ev
+	}
+
+	for i, ev := range evis {
+		if err := ev.ValidateBasic(); err != nil {
+			return nil, fmt.Errorf("invalid evidence (#%d): %v", i, err)
+		}
+	}
+
+	return evis, nil
 }
